requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
langchain==0.3.3
openai==1.51.2
requests==2.32.3
langchain-openai==0.2.2
python-dotenv==1.0.1

